refactor(utils): added util modules
