+++ src/components/player.js
@@ -1,0 +1,17 @@
+const buildPlayer = (song) => {
+    if (song.hasOwnProperty('source')) {
+        const sc_el =
+            `<iframe width="640" height="300" scrolling="no" frameborder="no" allow="autoplay" src="https://w.soundcloud.com/player/?url=https%3A//api.soundcloud.com/tracks/${song.trackId}&color=%23ff5500&auto_play=false&hide_related=false&show_comments=true&show_user=true&show_reposts=false&show_teaser=true&visual=true"></iframe><div style="font-size: 10px; color: #cccccc;line-break: anywhere;word-break: normal;overflow: hidden;white-space: nowrap;text-overflow: ellipsis; font-family: Interstate,Lucida Grande,Lucida Sans Unicode,Lucida Sans,Garuda,Verdana,Tahoma,sans-serif;font-weight: 100;"><a href="https://soundcloud.com/${song.link_author}" title="${song.author}" target="_blank" style="color: #cccccc; text-decoration: none;">${song.author}</a> · <a href="https://soundcloud.com/${song.link}" title="${song.name}" target="_blank" style="color: #cccccc; text-decoration: none;">${song.name}</a></div>`;
+
+        return sc_el;
+    }
+
+    const el =
+        `<iframe id="ytplayer" type="text/html" width="640" height="360"
+        src="https://www.youtube.com/embed/${song.link}?autoplay=0"
+        frameborder="0"></iframe>`;
+
+    return el;
+}
+
+export default buildPlayer;
