+++ src/utils/cookies.js
@@ -1,0 +1,22 @@
+export const createCookie = (name, value) => {
+    let currentDate = new Date();
+    let expirationDate = new Date(currentDate.getFullYear(), currentDate.getMonth(), currentDate.getDate() + 1, 0, 0, 0);
+    let expires = "; expires=" + expirationDate.toGMTString();
+    document.cookie = name + "=" + value + expires + "; path=/";
+}
+
+export const getCookie = (cname) => {
+    const name = cname + "=";
+    const decodedCookie = decodeURIComponent(document.cookie);
+    const ca = decodedCookie.split(';');
+    for (let i = 0; i < ca.length; i++) {
+        let c = ca[i];
+        while (c.charAt(0) === ' ') {
+            c = c.substring(1);
+        }
+        if (c.indexOf(name) === 0) {
+            return c.substring(name.length, c.length);
+        }
+    }
+    return "";
+}
